// apps/api/src/observation/enums/observation-category.enum.ts
export enum ObservationCategory {
    POWER = 'POWER',
    WATER = 'WATER',
    INTERNET = 'INTERNET',
    FLOODING = 'FLOODING',
    ROAD_ACCESS = 'ROAD_ACCESS',
    DRAINAGE = 'DRAINAGE',
    SECURITY = 'SECURITY',
    WASTE = 'WASTE',
    NOISE = 'NOISE',
    BUILDING_CONDITION = 'BUILDING_CONDITION',
    PARKING = 'PARKING',
    SANITATION = 'SANITATION',
    LIGHTING = 'LIGHTING',
    ENVIRONMENT = 'ENVIRONMENT',
    CONSTRUCTION = 'CONSTRUCTION',
    OTHER = 'OTHER',
  }